create folder1 and add move all file to folder1
